Fixed handling of invalid values. Added NaN constant. Force copy of data when assigning to a symbol so that s1 = s2 = expr does the right thing (s1 and s2 should be separate copies of the data). Updated man page.

+++ progs/minccalc/node.h
@@ -80,7 +80,7 @@ struct node {
    node_t expr[3];
    ident_t ident;
    int   flags;
-   float real;
+   double real;
    int   pos;
    int   numargs;
 };
